+++ celer/__init__.py
@@ -2,6 +2,6 @@
 
 from .homotopy import celer_path
 from .wrapper import celer
-from .dropin_sklearn import Lasso, LassoCV
+from .dropin_sklearn import Lasso, LassoCV, LogisticRegression
 
 __version__ = '0.5dev'