+++ celer/tests/test_lasso.py
@@ -13,10 +13,11 @@ from sklearn.linear_model._logistic import _logistic_regression_path
 from sklearn.exceptions import ConvergenceWarning
 from sklearn.utils.estimator_checks import check_estimator
 from sklearn.linear_model import (LassoCV as sklearn_LassoCV,
-                                  Lasso as sklearn_Lasso, lasso_path)
+                                  Lasso as sklearn_Lasso, lasso_path,
+                                  LogisticRegression as sklearn_Logreg)
 
 from celer import celer_path, celer
-from celer.dropin_sklearn import Lasso, LassoCV
+from celer.dropin_sklearn import Lasso, LassoCV, LogisticRegression
 from celer.utils.testing import build_dataset
 
 
@@ -37,6 +38,35 @@ def test_logreg():
     np.testing.assert_array_less(gaps, tol)
     np.testing.assert_allclose(coefs != 0, coefs_c.T != 0)
     np.testing.assert_allclose(coefs, coefs_c.T, atol=1e-5, rtol=1e-3)
+
+
+def test_dropin_logreg():
+    np.random.seed(1409)
+    check_estimator(LogisticRegression)
+    X, y, _, _ = build_dataset(
+        n_samples=100, n_features=100, sparse_X=True)
+    y = np.sign(y)
+    alpha_max = norm(X.T.dot(y), ord=np.inf) / 2
+    C = 30. / alpha_max
+
+    tol = 1e-8
+    clf1 = LogisticRegression(C=C, tol=tol)
+    clf1.fit(X, y)
+
+    clf2 = sklearn_Logreg(
+        C=C, penalty='l1', solver='liblinear', fit_intercept=False, tol=tol)
+    clf2.fit(X, y)
+    np.testing.assert_allclose(clf1.coef_, clf2.coef_, rtol=1e-3, atol=1e-5)
+
+    # multinomial test:
+    y = np.random.choice(4, len(y))
+    clf3 = LogisticRegression(C=C, tol=tol)
+    clf3.fit(X, y)
+
+    clf4 = sklearn_Logreg(
+        C=C, penalty='l1', solver='liblinear', fit_intercept=False, tol=tol)
+    clf4.fit(X, y)
+    np.testing.assert_allclose(clf3.coef_, clf4.coef_, rtol=1e-3, atol=1e-4)
 
 
 @pytest.mark.parametrize("sparse_X, alphas, pb",
@@ -191,3 +221,7 @@ def test_warm_start():
         reg1.fit(X, y)
         # hack because assert_array_less does strict comparison...
         np.testing.assert_array_less(reg1.n_iter_, 2.01)
+
+
+if __name__ == "__main__":
+    test_dropin_logreg()