+++ whats_new.rst
@@ -3,8 +3,9 @@ Version 0.5
 
 Changelog
 ~~~~~~~~~
-- Added a datasets module, now supporting more datasets (climate for Sparse Group Lasso)
+- Added module `datasets`, supporting more datasets (climate for Sparse Group Lasso)
 - Removed `celer_logreg` function, use `celer` instead with `pb="logreg"`
+- Added sklearn-like `LogisticRegression` class.
 
 Version 0.4
 ------------
@@ -19,4 +20,4 @@ Version 0.3.1
 
 Changelog
 ~~~~~~~~~
-- Fixed bugs in screenings.
+- Fixed bugs in screening.